requirements.txt
streamlit
matplotlib
opencv-python
pinecone
scipy
mtcnn
keras-facenet
python-dotenv
scikit-learn